backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
firebase-admin==6.4.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
google-generativeai==0.3.2
requests==2.31.0
moviepy==1.0.3
Pillow>=9.0.0
google-auth>=2.26.1
imageio-ffmpeg==0.4.9